# zstd: Improve "best" compression