@klauspost 2022-09-25
Very nice! Could you add a comment explaining how you determined the size?

That way if the struct is changed whoever is looking at it will know how to adjust it.